[test] moves the trigger event to its own helper
