+++ tests/integration/components/ember-kalendae-test.js
@@ -1,43 +1,10 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
+import TriggerEvent from '../../helpers/ember-kalendae/trigger-event';
 
 moduleForComponent('ember-kalendae', 'Integration | Component | ember kalendae', {
   integration: true
 });
-
-// taken from http://jsfiddle.net/mendesjuan/rHMCy/4/
-const fireEvent = function(node, eventName) {
-   if (node.dispatchEvent) {
-    // Gecko-style approach (now the standard) takes more work
-    let eventClass = "";
-
-    // Different events have different event classes.
-    // If this switch statement can't map an eventName to an eventClass,
-    // the event firing is going to fail.
-    switch (eventName) {
-      case 'click': // Dispatching of 'click' appears to not work correctly in Safari. Use 'mousedown' or 'mouseup' instead.
-      case 'mousedown':
-      case 'mouseup':
-        eventClass = 'MouseEvents';
-        break;
-
-      case 'focus':
-      case 'change':
-      case 'blur':
-      case 'select':
-        eventClass = 'HTMLEvents';
-        break;
-
-      default:
-        throw `fireEvent: Couldn't find an event class for event '${eventName}.`;
-    }
-    var event = node.ownerDocument.createEvent(eventClass);
-    event.initEvent(eventName, true, true); // All events created as bubbling and cancelable.
-    event.synthetic = true; // allow detection of synthetic events
-    // The second parameter says go ahead with the default action
-    node.dispatchEvent(event, true);
-  }
-};
 
 test('it allows block usage', function(assert) {
   assert.expect(1);
@@ -75,7 +42,7 @@ test('it triggers date changed event correctly and passes the selected date thro
   });
 
   this.render(hbs`{{ember-kalendae onDateClicked=(action onDateClicked)}}`);
-  fireEvent(this.$().find('.k-active:first')[0], 'mousedown');
+  TriggerEvent(this.$().find('.k-active:first')[0], 'mousedown');
 });
 
 test('it passes mode through to kalendae correctly', function(assert) {
